split up cell and world files and specs
