+++ spec/spec_helper.rb
@@ -4,9 +4,17 @@
 # loaded once.
 #
 # See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
+
+Dir.glob("./spec/support/**/*.rb").each { |f| require f }
+
 RSpec.configure do |config|
   config.run_all_when_everything_filtered = true
   config.filter_run :focus
+
+  config.include Helpers
+
+  config.color = true
+  config.formatter = "documentation"
 
   # Run specs in random order to surface order dependencies. If you find an
   # order dependency and want to debug it, you can fix the order by providing
